refactor: 응원 모달 분리
