+++ features/record-detail/components/index.ts
@@ -1,5 +1,3 @@
-export * from './cheer-bottom-sheet';
-export * from './cheer-item';
 export * from './cheer-modal';
 export * from './cheer-progress';
 export * from './date-picker';
